refactored the api, to be in the ACTUAL api map :tada:
