+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,33 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="20">
+            <item index="0" class="java.lang.String" itemvalue="typing_extensions" />
+            <item index="1" class="java.lang.String" itemvalue="torch" />
+            <item index="2" class="java.lang.String" itemvalue="tqdm" />
+            <item index="3" class="java.lang.String" itemvalue="pygame" />
+            <item index="4" class="java.lang.String" itemvalue="matplotlib-inline" />
+            <item index="5" class="java.lang.String" itemvalue="numpydoc" />
+            <item index="6" class="java.lang.String" itemvalue="gym-notices" />
+            <item index="7" class="java.lang.String" itemvalue="opencv-python" />
+            <item index="8" class="java.lang.String" itemvalue="matplotlib" />
+            <item index="9" class="java.lang.String" itemvalue="numpy" />
+            <item index="10" class="java.lang.String" itemvalue="torchvision" />
+            <item index="11" class="java.lang.String" itemvalue="torchaudio" />
+            <item index="12" class="java.lang.String" itemvalue="Flask" />
+            <item index="13" class="java.lang.String" itemvalue="fsspec" />
+            <item index="14" class="java.lang.String" itemvalue="fastapi" />
+            <item index="15" class="java.lang.String" itemvalue="networkx" />
+            <item index="16" class="java.lang.String" itemvalue="filelock" />
+            <item index="17" class="java.lang.String" itemvalue="pillow" />
+            <item index="18" class="java.lang.String" itemvalue="MarkupSafe" />
+            <item index="19" class="java.lang.String" itemvalue="Jinja2" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
