+++ backend/api/models/requests.py
@@ -1,10 +1,18 @@
 from pydantic import BaseModel
 
+
 class AIGameRequest(BaseModel):
     white_model: str
     black_model: str
+
 
 class MoveRequest(BaseModel):
     move: str
     turn: int
     board: list
+
+
+class ActionRequest(BaseModel):
+    board: list
+    turn: int
+    pieceLocation: str
