+++ backend/api/models/responses.py
@@ -1,16 +1,20 @@
-import numpy as np
 from pydantic import BaseModel
+
 
 class AIGameResponse(BaseModel):
     game: list
     statistics: list
 
+
 class MoveResponse(BaseModel):
-    board: list
+    playerMoveBoard: list
+    CombinedMoveBoard: list
     cards: list
     resources: int
     has_game_ended: bool
 
+class ActionResponse(BaseModel):
+    possibleMoves: list
 
 class InitializeResponse(BaseModel):
     board: list
