+++ frontend/nuxt.config.ts
