+++ frontend/pages/play.vue
