+++ frontend/pages/playAI.vue
